Experimenting with scaffold

+++ Rakefile
@@ -12,6 +12,8 @@ task :generate_erd do
   system "rails db:migrate"
   system "EAGER_LOAD=1 bundle exec erd --inheritance --filetype=dot --direct --attributes=foreign_keys,content"
   system "dot -Tpng erd.dot > ./db/erd/generated/erd_#{DateTime.now.strftime('%Q')}.png"
-  File.delete('erd.dot')
-  File.delete('erd.pdf')
+  File.delete('erd.dot') if File.exist?('erd.dot') if File.exist?('erd.dot')
+  File.delete('erd.pdf') if File.exist?('erd.pdf') if File.exist?('erd.pdf')
+  system "rake db:seed"
+  system "rails s"
 end
